server/pyproject.toml
[project]
name = "gmail-mcp"
version = "0.1.0"
description = "Multi-account Gmail MCP server with hardened auth and draft-confirm sending"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "keyring>=24.0.0",
    "html2text>=2024.2.26",
]

[project.scripts]
gmail-mcp = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
